backend/requirements.txt
# ===================================================================
# REQUIREMENTS.TXT CONSOLIDÉ POUR BUDGET FAMILLE v2.3 - UBUNTU WSL
# ===================================================================
# Configuration optimisée pour Ubuntu/WSL avec Python 3.8.10
# Consolidation de tous les requirements fragmentés

# === CORE API FRAMEWORK ===
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.31.0

# === DATA PROCESSING ===
# Versions compatibles Python 3.8.10 Ubuntu
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.26.0

# === DATABASE & ORM ===
SQLAlchemy>=1.4.0,<2.1.0
alembic>=1.8.0  # Pour migrations de DB

# === SECURITY & AUTHENTICATION ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=40.0.0

# === HTTP & VALIDATION ===
python-multipart>=0.0.5
pydantic>=1.10.0,<2.6.0
email-validator>=1.3.0

# === FILE HANDLING ===
# Détection MIME optimisée pour Ubuntu
python-magic>=0.4.20
openpyxl>=3.0.0  # Pour Excel

# === CONFIGURATION & LOGGING ===
python-dotenv>=0.19.0
structlog>=22.0.0  # Logging structuré

# === OPTIONAL - Database Encryption ===
# Décommenté si chiffrement DB requis (nécessite libsqlcipher-dev)
# pysqlcipher3>=1.2.0

# ===================================================================
# INSTALLATION UBUNTU WSL:
# 
# 1. Mise à jour système:
# sudo apt update && sudo apt upgrade -y
# 
# 2. Dépendances système:
# sudo apt install -y libmagic1 libmagic-dev python3-dev
# 
# 3. Installation Python:
# pip install -r requirements.txt
# 
# 4. Pour chiffrement DB (optionnel):
# sudo apt install -y libsqlcipher-dev
# pip install pysqlcipher3
# 
# 5. Pour tests de développement:
# pip install pytest pytest-asyncio httpx
# ===================================================================